Rename the new header to just game.h
