+++ src/game_logic_classic/vars.h
@@ -25,7 +25,7 @@
 
 #pragma once
 
-#include "gamedefs.h"
+#include "game.h"
 
 
 extern bool sysTecMode;
